PaleMossBlock
