+++ tests/phpstan/configs/actual-problems.neon
@@ -357,19 +357,37 @@ parameters:
 		-
 			message: '#^Parameter \#1 \$x of method pocketmine\\world\\World\:\:getBlockAt\(\) expects int, float\|int given\.$#'
 			identifier: argument.type
-			count: 5
+			count: 3
+			path: ../../../src/block/MossBlock.php
+
+		-
+			message: '#^Parameter \#1 \$x of method pocketmine\\world\\World\:\:isInWorld\(\) expects int, float\|int given\.$#'
+			identifier: argument.type
+			count: 2
 			path: ../../../src/block/MossBlock.php
 
 		-
 			message: '#^Parameter \#2 \$y of method pocketmine\\world\\World\:\:getBlockAt\(\) expects int, float\|int given\.$#'
 			identifier: argument.type
-			count: 5
+			count: 3
+			path: ../../../src/block/MossBlock.php
+
+		-
+			message: '#^Parameter \#2 \$y of method pocketmine\\world\\World\:\:isInWorld\(\) expects int, float\|int given\.$#'
+			identifier: argument.type
+			count: 2
 			path: ../../../src/block/MossBlock.php
 
 		-
 			message: '#^Parameter \#3 \$z of method pocketmine\\world\\World\:\:getBlockAt\(\) expects int, float\|int given\.$#'
 			identifier: argument.type
-			count: 5
+			count: 3
+			path: ../../../src/block/MossBlock.php
+
+		-
+			message: '#^Parameter \#3 \$z of method pocketmine\\world\\World\:\:isInWorld\(\) expects int, float\|int given\.$#'
+			identifier: argument.type
+			count: 2
 			path: ../../../src/block/MossBlock.php
 
 		-
@@ -401,6 +419,24 @@ parameters:
 			identifier: argument.type
 			count: 1
 			path: ../../../src/block/Nylium.php
+
+		-
+			message: '#^Parameter \#1 \$x of method pocketmine\\world\\World\:\:isInWorld\(\) expects int, float\|int given\.$#'
+			identifier: argument.type
+			count: 2
+			path: ../../../src/block/PaleMossBlock.php
+
+		-
+			message: '#^Parameter \#2 \$y of method pocketmine\\world\\World\:\:isInWorld\(\) expects int, float\|int given\.$#'
+			identifier: argument.type
+			count: 2
+			path: ../../../src/block/PaleMossBlock.php
+
+		-
+			message: '#^Parameter \#3 \$z of method pocketmine\\world\\World\:\:isInWorld\(\) expects int, float\|int given\.$#'
+			identifier: argument.type
+			count: 2
+			path: ../../../src/block/PaleMossBlock.php
 
 		-
 			message: '#^Parameter \#1 \$x of method pocketmine\\world\\World\:\:getFullLightAt\(\) expects int, float\|int given\.$#'
